Fix animation can see through problem
We found that for the animations that uses paint worklet infra,
which are background-color and custom property animations, we can
see through the browser window when there is one keyframe with
alpha < 1.

Investigation shows that the reason is that the contents_opaque_
in the LayerImpl isn't correctly set. This CL fixes the problem
and add tests.

Bug: 1244155, 1237234
Change-Id: Iecddfa2c335f65e1857a96ae7790eaa4e91952d0