Don't adjust untrusted TouchEvents.
Untrusted touch events are not fired to DOM for whatever reason.  But
they still go through the adjustment and causes problem when event
target is an empty document.

We need to fix wpt/touch-events/touch-retargeting.html by avoiding
untrusted event dispatch but test_driver doesn't support input to shadow
DOM yet.  So switched the test to be manual for now.  I manually
verified that both Chrome and Firefox passes the manual test.

Bug: 1230583
Change-Id: Ia82cbf4a65d859d76180dae063ee7840efb7f3ab