[WPT/referrer-policy] Use unsafe-url as "another policy"
Previously, `no-referrer` was used as the Document's referrer policy
that shouldn't be used.
But using the Document's referrer policy unexpectedly didn't cause
test failures, because
https://github.com/web-platform-tests/wpt/pull/29434 made
the tests to allow UA-specific policies more strict than expected,
and using `no-referrer` is always allowed.

This CL uses `unsafe-url` instead as long as possible, to make
tests fail if the policy is used unexpectedly.
`unsafe-url` is likely to result in less strict referrers that
aren't allowed as UA-specific policies.

Bug: 1235205, https://github.com/web-platform-tests/wpt/pull/29434
Change-Id: I7e9aa4f5e5fbd4d8ff5a984f6647845ae0d0e2f1